# docs/swagger.yaml
definitions:
  board_columns_dtos.BoardColumnsRequest:
    properties:
      name:
        type: string
      position:
        type: integer
      workspace_id:
        type: integer
    type: object
...
  models.TwBoardColumn:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      position:
        type: integer
      updated_at:
        type: string
      workspace:
        $ref: '#/definitions/models.TwWorkspace'
      workspace_id:
        type: integer
    type: object
...
  models.TwDocument:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      download_url:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      schedule:
        $ref: '#/definitions/models.TwSchedule'
      schedule_id:
        type: integer
      updated_at:
        type: string
      uploaded_at:
        type: string
      workspace_user_id:
        type: integer
      workspaceUser:
        $ref: '#/definitions/models.TwWorkspaceUser'
    type: object
  models.TwSchedule:
    properties:
      all_day:
        type: boolean
      board_column_id:
        type: integer
      boardColumn:
        $ref: '#/definitions/models.TwBoardColumn'
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      end_time:
        type: string
      extra_data:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      location:
        type: string
      recurrence_pattern:
        type: string
      start_time:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      video_transcript:
        type: string
      visibility:
        type: string
      workspace:
        $ref: '#/definitions/models.TwWorkspace'
      workspace_id:
        type: integer
      workspace_user_id:
        type: integer
      workspaceUser:
        $ref: '#/definitions/models.TwWorkspaceUser'
    type: object
...
  models.TwUser:
    properties:
      calendarSettings:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      firstName:
        type: string
      googleId:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isVerified:
        type: boolean
      lastLoginAt:
        type: string
      lastName:
        type: string
      locale:
        type: string
      notificationSettings:
        type: string
      profilePicture:
        type: string
      role:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
    type: object
  models.TwUserEmail:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.TwUser'
      user_id:
        type: integer
    type: object
  models.TwWorkspace:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      extra_data:
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      key:
        type: string
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  models.TwWorkspaceUser:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      extra_data:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_verified:
        type: boolean
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_email_id:
        type: integer
      userEmail:
        $ref: '#/definitions/models.TwUserEmail'
      workspace:
        $ref: '#/definitions/models.TwWorkspace'
      workspace_id:
        type: integer
      workspace_key:
        type: string
    type: object
...
info:
  contact: {}
  description: Timewise database management system
  title: timewise-dbms
  version: "1.0"
paths:
  /dbms/v1/board_columns:
    post:
      consumes:
      - application/json
      description: Create board column
      parameters:
      - description: Create board column request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/board_columns_dtos.BoardColumnsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TwBoardColumn'
      summary: Create board column
      tags:
      - board_columns
  /dbms/v1/board_columns/{id}:
    delete:
      consumes:
      - application/json
      description: Delete board column
      parameters:
      - description: Board column ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete board column
      tags:
      - board_columns
    get:
      consumes:
      - application/json
      description: Get board column by ID
      parameters:
      - description: Board column ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TwBoardColumn'
      summary: Get board column by ID
      tags:
      - board_columns
    put:
      consumes:
      - application/json
      description: Update board column
      parameters:
      - description: Board column ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update board column request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TwBoardColumn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TwBoardColumn'
      summary: Update board column
      tags:
      - board_columns
  /dbms/v1/board_columns/{id}/{field}:
    get:
      consumes:
      - application/json
      description: Get board column field
      parameters:
      - description: Board column ID
        in: path
        name: id
        required: true
        type: integer
      - description: Field
        in: path
        name: field
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TwBoardColumn'
      summary: Get board column field
      tags:
      - board_columns
    put:
      consumes:
      - application/json
      description: Update board column field
      parameters:
      - description: Board column ID
        in: path
        name: id
        required: true
        type: integer
      - description: Field
        in: path
        name: field
        required: true
        type: string
      - description: Update board column request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TwBoardColumn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TwBoardColumn'
      summary: Update board column field
      tags:
      - board_columns
...
  /dbms/v1/document/schedule/{schedule_id}:
    get:
      consumes:
      - application/json
      description: Get documents by schedule
      parameters:
      - description: Schedule ID
        in: path
        name: schedule_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TwDocument'
            type: array
      summary: Get documents by schedule
      tags:
      - document
...
  /dbms/v1/workspace/{workspace_id}/board_columns:
    get:
      consumes:
      - application/json
      description: Get board columns by workspace
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TwBoardColumn'
            type: array
      summary: Get board columns by workspace
      tags:
      - board_columns
...
swagger: "2.0"